Maintenance Update: 1

+++ components/Debugger/src/core/benchmark/profiler.cpp
@@ -89,7 +89,7 @@ namespace Dynamik {
 				auto stop = std::chrono::time_point_cast<std::chrono::microseconds>
 					(endTimePoint).time_since_epoch().count();
 
-				uint32_t threadID = std::hash<std::thread::id>{}(std::this_thread::get_id());
+				uint32_t threadID = (uint32_t)std::hash<std::thread::id>{}(std::this_thread::get_id());
 				benchmarkResult results = {};
 				results.name = myName;
 				results.start = start;
